# Improve `TwoBodyElectronicIntegrals.to_second_q_op` performance