--- a/qiskit_nature/properties/second_quantization/electronic/integrals/electronic_integrals.py
+++ b/qiskit_nature/properties/second_quantization/electronic/integrals/electronic_integrals.py
@@ -15,7 +15,6 @@
 from __future__ import annotations
 
 import importlib
-import itertools
 from abc import ABC, abstractmethod
 from copy import deepcopy
 from typing import Generator, Optional, Union