--- a/qiskit_nature/properties/second_quantization/electronic/integrals/two_body_electronic_integrals.py
+++ b/qiskit_nature/properties/second_quantization/electronic/integrals/two_body_electronic_integrals.py
@@ -14,12 +14,15 @@
 
 from __future__ import annotations
 
+import itertools
 from typing import Optional, Union, cast
 
 import numpy as np
 
+from qiskit.tools import parallel_map
 from qiskit_nature import QiskitNatureError
 from qiskit_nature.settings import settings
+from qiskit_nature.operators.second_quantization import FermionicOp
 
 from .electronic_integrals import ElectronicIntegrals
 from .one_body_electronic_integrals import OneBodyElectronicIntegrals
@@ -181,8 +184,9 @@ class TwoBodyElectronicIntegrals(ElectronicIntegrals):
 
         return np.where(np.abs(so_matrix) > self._threshold, so_matrix, 0.0)
 
-    def _calc_coeffs_with_ops(self, indices: tuple[int, ...]) -> list[tuple[int, str]]:
-        return [(indices[0], "+"), (indices[2], "+"), (indices[3], "-"), (indices[1], "-")]
+    @staticmethod
+    def _calc_coeffs_with_ops(indices: tuple[int, ...]) -> list[tuple[str, int]]:
+        return [("+", indices[0]), ("+", indices[2]), ("-", indices[3]), ("-", indices[1])]
 
     def compose(
         self, other: ElectronicIntegrals, einsum_subscript: Optional[str] = None