--- a/releasenotes/notes/performance-two-body-second-q-op-08f906951005ebc1.yaml
+++ b/releasenotes/notes/performance-two-body-second-q-op-08f906951005ebc1.yaml
@@ -1,0 +1,9 @@
+---
+features:
+  - |
+    Improves the performance of the
+    :class:`qiskit_nature.properties.second_quantization.electronic.integrals.TwoBodyElectronicIntegrals.to_second_q_op`
+    method significantly. Previously, generating the second quantized operators
+    for a system size greater than 20 spin orbitals took on the order of minutes
+    to hours (depending on the actual size). Now, even system sizes of 50 spin
+    orbitals can be handled in a matter of seconds.