@woodsp-ibm 2022-04-13
So basically this builds out the FermionicOp having first created all the terms rather than building one and then augmenting it term by term. Is there any aspect of the performance around how the methods used before that could considered in the future though perhaps always building it out in one go may faster than how things were done.