@mrossinek 2022-04-14
Yes I believe that the previous way of building out the `FermionicOp` will always be less performant, simply because of the overhead involved with constructed 5 instances of `FermionicOp` per iteration, compared to a single instantiation in the new case. Obviously, this single instantiation now can take considerably longer than the previous single ones, but as you can see from the speedup achieved here, this is not a problem.

I think the old implementation was a bit more instructive/educational in terms of how the integral indices determine which creation/annihilation operator to use, but this is not worth the performance deficiencies.